Algoritmos/Projeto Arvore Binaria/src/python/requirements.txt
pandas>=2.0
openpyxl>=3.0
psutil>=5.0
py-cpuinfo>=9.0
